requirements.txt
faiss-cpu
numpy
tqdm
PyMuPDF
beautifulsoup4
requests
sentence-transformers
ollama